// gme-web/gme-web/src/app/@data/model/gestores-orh.ts
export class GestoresOrh {
  gestorId?: number;
  entidadId?: number;
  personaId?: number;
  tipoDocumento?: number;
  numeroDocumento?: string;
  nombres?: string;
  apellidoPaterno?: string;
  apellidoMaterno?: string;
  nombreCompleto?: string;
  correo?: string;
  numeroTelefono?: string;
  fechaNacimiento?: string;
  usuarioId?: number;
  rolId?: number;
  estadoRegistro?: string;
  estado?: string;
  colorEstado?: string;
  disabledNotification?: boolean;
  disabledDelete?: boolean;
  TelefonoId?: number;
  numeroCelular?: string;
  celularId?: number;
}